testsuite: parser: integer: re_tree_integer: new
If we break out the regexes and testing commands from one another, that
allows the user to pass a regex for one tree type as an argument into
another.  This allows a tree of regex's to be built that should match
the parse tree.

Add a new function for building a tree integer regex and using that in
the expect_tree_integer function.