x86_64: function_call: int_args: new
Add a new list sub-list of arguments that should be placed in the
integer registers.  This approach will allow us to interface with the
assembly code by constructing a list in arg_head that will have a
specific order, thereby allowing the assembly code to move parameters
into the correct register.