tree: make live_var's val a pointer
So we can create sets of live variables that are contiguous in
memory (when declaring arrays, structs, etc), we have to be able to
control the location of the live_var's value.  The first step is to
make the val member a pointer so we can control where it is located.