gc: fix a bug in `mark_stack'
Move __builtin_unwind_init() to collect() so that all callee-saved
registers are on the stack before mark_stack() is called.  If
__builtin_unwind_init() is in the same function that is marking the
stack, there is a possibility that the stack slot for `bottom_of_stack'
would be above the callee-saved registers and the objects would not be
marked.

Also mark mark_stack() as noinline to prevent the compiler from inlining
it with collect(), preventing the same problem as above.