repl: fix case where preprocessed source has single line
If the source file that we're reading in after preprocessing has been
ran only contains one line then the code to read this is gets stuck in
an infinte loop.

Fix this by reading back until a '\n' OR the beginning of the file.
Also provide a simple testcase for this.